Move P2pConnectionContext to separate files
